requirements/production.txt
setuptools>=35.0.2
google-cloud-datastore==1.0.0
google-cloud-storage==1.1.0
google-cloud-pubsub==0.25.0
psutil==5.2.2
requests==2.13.0
lxml==3.7.3
requests-file==1.4.2
python-dateutil==2.6.0
click==6.7
